asynq_server

Module message

WebSocket message types for asynq-server protocol

This module defines the message protocol used for WebSocket communication between clients and the asynq-server.

Structs

AddToGroupRequest

Request to add a task to a group for aggregation

AddToGroupUniqueRequest

Request to add a unique task to a group for aggregation

AggregationCheckRequest

Request to check aggregation conditions

ArchiveRequest

Request to archive a task

DeleteAggregationSetRequest

Request to delete an aggregation set

DequeueRequest

Request to dequeue tasks

EnqueueRequest

Request to enqueue a task

EnqueueUniqueRequest

Request to enqueue a unique task

ListGroupsRequest

Request to list groups in a queue

ReadAggregationSetRequest

Request to read tasks from an aggregation set

RetryRequest

Request to retry a task

ScheduleRequest

Request to schedule a task

ScheduleUniqueRequest

Request to schedule a unique task

TaskDoneRequest

Request to mark a task as done or complete

TaskInfoResponse

Task info response

TaskMessageResponse

Task message response (for dequeue)

Enums

ClientMessage

Request message from client to server

ServerMessage

Response message from server to client